Prepare to get rid of old config logic
